refactor(transformer): move computed key methods to common/computed_keys
