+++ crates/oxc_transformer/src/es2022/class_properties/computed_key.rs
@@ -3,10 +3,9 @@
 
 use oxc_allocator::TakeIn;
 use oxc_ast::ast::*;
-use oxc_syntax::symbol::SymbolFlags;
 use oxc_traverse::TraverseCtx;
 
-use super::{ClassProperties, utils::create_assignment};
+use super::ClassProperties;
 
 impl<'a> ClassProperties<'a, '_> {
     /// Substitute temp var for method computed key.
@@ -25,7 +24,7 @@ impl<'a> ClassProperties<'a, '_> {
 
         // Exit if evaluating key cannot have side effects.
         // This check also results in exit for non-computed keys e.g. `class C { 'x'() {} 123() {} }`.
-        if !key_needs_temp_var(key, ctx) {
+        if !self.ctx.key_needs_temp_var(key, ctx) {
             return;
         }
 
@@ -36,7 +35,7 @@ impl<'a> ClassProperties<'a, '_> {
         //    or class contains a static block which is being transformed
         //    (static blocks are always evaluated after computed keys, regardless of order)
         let original_key = key.take_in(ctx.ast.allocator);
-        let (assignment, temp_var) = self.create_computed_key_temp_var(original_key, ctx);
+        let (assignment, temp_var) = self.ctx.create_computed_key_temp_var(original_key, ctx);
         self.insert_before.push(assignment);
         method.key = PropertyKey::from(temp_var);
     }
@@ -64,8 +63,8 @@ impl<'a> ClassProperties<'a, '_> {
         ctx: &mut TraverseCtx<'a>,
     ) -> Expression<'a> {
         let original_key = key.take_in(ctx.ast.allocator);
-        if key_needs_temp_var(&original_key, ctx) {
-            let (assignment, ident) = self.create_computed_key_temp_var(original_key, ctx);
+        if self.ctx.key_needs_temp_var(&original_key, ctx) {
+            let (assignment, ident) = self.ctx.create_computed_key_temp_var(original_key, ctx);
             if is_static {
                 self.insert_before.push(assignment);
             } else {
@@ -75,26 +74,6 @@ impl<'a> ClassProperties<'a, '_> {
         } else {
             original_key
         }
-    }
-
-    /// Create `let _x;` statement and insert it.
-    /// Return `_x = x()` assignment, and `_x` identifier referencing same temp var.
-    fn create_computed_key_temp_var(
-        &self,
-        key: Expression<'a>,
-        ctx: &mut TraverseCtx<'a>,
-    ) -> (/* assignment */ Expression<'a>, /* identifier */ Expression<'a>) {
-        let outer_scope_id = ctx.current_block_scope_id();
-        // TODO: Handle if is a class expression defined in a function's params.
-        let binding =
-            ctx.generate_uid_based_on_node(&key, outer_scope_id, SymbolFlags::BlockScopedVariable);
-
-        self.ctx.var_declarations.insert_let(&binding, None, ctx);
-
-        let assignment = create_assignment(&binding, key, ctx);
-        let ident = binding.create_read_expression(ctx);
-
-        (assignment, ident)
     }
 
     /// Extract computed key if it's an assignment, and replace with identifier.
@@ -140,54 +119,5 @@ impl<'a> ClassProperties<'a, '_> {
         // Extract assignment from computed key and insert before class
         let assignment = prop.key.take_in(ctx.ast.allocator).into_expression();
         self.insert_before.push(assignment);
-    }
-}
-
-/// Check if temp var is required for `key`.
-///
-/// `this` does not have side effects, but in this context, it needs a temp var anyway, because `this`
-/// in computed key and `this` within class constructor resolve to different `this` bindings.
-/// So we need to create a temp var outside of the class to get the correct `this`.
-/// `class C { [this] = 1; }`
-/// -> `let _this; _this = this; class C { constructor() { this[_this] = 1; } }`
-//
-// TODO(improve-on-babel): Can avoid the temp var if key is for a static prop/method,
-// as in that case the usage of `this` stays outside the class.
-fn key_needs_temp_var(key: &Expression, ctx: &TraverseCtx) -> bool {
-    match key {
-        // Literals cannot have side effects.
-        // e.g. `let x = 'x'; class C { [x] = 1; }` or `class C { ['x'] = 1; }`.
-        Expression::BooleanLiteral(_)
-        | Expression::NullLiteral(_)
-        | Expression::NumericLiteral(_)
-        | Expression::BigIntLiteral(_)
-        | Expression::RegExpLiteral(_)
-        | Expression::StringLiteral(_) => false,
-        // Template literal cannot have side effects if it has no expressions.
-        // If it *does* have expressions, but they're all literals, then also cannot have side effects,
-        // but don't bother checking for that as it shouldn't occur in real world code.
-        // Why would you write "`x${9}z`" when you can just write "`x9z`"?
-        // Note: "`x${foo}`" *can* have side effects if `foo` is an object with a `toString` method.
-        Expression::TemplateLiteral(lit) => !lit.expressions.is_empty(),
-        // `IdentifierReference`s can have side effects if is unbound.
-        //
-        // If var is mutated, it also needs a temp var, because of cases like
-        // `let x = 1; class { [x] = 1; [++x] = 2; }`
-        // `++x` is hoisted to before class in output, so `x` in 1st key would get the wrong value
-        // unless it's hoisted out too.
-        //
-        // TODO: Add an exec test for this odd case.
-        // TODO(improve-on-babel): That case is rare.
-        // Test for it in first pass over class elements, and avoid temp vars where possible.
-        Expression::Identifier(ident) => {
-            match ctx.scoping().get_reference(ident.reference_id()).symbol_id() {
-                Some(symbol_id) => ctx.scoping().symbol_is_mutated(symbol_id),
-                None => true,
-            }
-        }
-        // Treat any other expression as possibly having side effects e.g. `foo()`.
-        // TODO: Do fuller analysis to detect expressions which cannot have side effects.
-        // e.g. `"x" + "y"`.
-        _ => true,
     }
 }
